[refactor] update template struct
