+++ tests/parts/utility_functions_test.go
@@ -1,406 +1,406 @@
 package parts
 
-import (
-	"go/format"
-	"go/parser"
-	"go/token"
-	"testing"
+// import (
+// 	"go/format"
+// 	"go/parser"
+// 	"go/token"
+// 	"testing"
 
-	"github.com/Mad-Pixels/go-dyno/internal/schema/common"
-	"github.com/Mad-Pixels/go-dyno/internal/utils"
-	v2 "github.com/Mad-Pixels/go-dyno/templates/v2"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
+// 	"github.com/Mad-Pixels/go-dyno/internal/schema/common"
+// 	"github.com/Mad-Pixels/go-dyno/internal/utils"
+// 	v2 "github.com/Mad-Pixels/go-dyno/templates/v2"
+// 	"github.com/stretchr/testify/assert"
+// 	"github.com/stretchr/testify/require"
+// )
 
-// TestUtilityFunctionsTemplate validates the UtilityFunctionsTemplate rendering.
-// This template generates a set of helper functions for DynamoDB operations:
-// - BatchPutItems, PutItem
-// - ExtractFromDynamoDBStreamEvent with per-type branches (including SS/NS)
-// - IsFieldModified, GetBoolFieldChanged
-// - CreateKey, CreateKeyFromItem
-// - CreateTriggerHandler
-// - ConvertMapToAttributeValues with Set support
-func TestUtilityFunctionsTemplate(t *testing.T) {
-	// Test with schema including all supported types including SS and NS
-	t.Run("basic_template_rendering", func(t *testing.T) {
-		templateMap := v2.TemplateMap{
-			PackageName: "usertable",
-			TableName:   "UserTable",
-			HashKey:     "user_id",
-			RangeKey:    "created_at",
-			AllAttributes: []common.Attribute{
-				{Name: "user_id", Type: "S"},
-				{Name: "created_at", Type: "N"},
-				{Name: "status", Type: "S"},
-				{Name: "updated_at", Type: "N"},
-				{Name: "is_active", Type: "BOOL"},
-			},
-		}
+// // TestUtilityFunctionsTemplate validates the UtilityFunctionsTemplate rendering.
+// // This template generates a set of helper functions for DynamoDB operations:
+// // - BatchPutItems, PutItem
+// // - ExtractFromDynamoDBStreamEvent with per-type branches (including SS/NS)
+// // - IsFieldModified, GetBoolFieldChanged
+// // - CreateKey, CreateKeyFromItem
+// // - CreateTriggerHandler
+// // - ConvertMapToAttributeValues with Set support
+// func TestUtilityFunctionsTemplate(t *testing.T) {
+// 	// Test with schema including all supported types including SS and NS
+// 	t.Run("basic_template_rendering", func(t *testing.T) {
+// 		templateMap := v2.TemplateMap{
+// 			PackageName: "usertable",
+// 			TableName:   "UserTable",
+// 			HashKey:     "user_id",
+// 			RangeKey:    "created_at",
+// 			AllAttributes: []common.Attribute{
+// 				{Name: "user_id", Type: "S"},
+// 				{Name: "created_at", Type: "N"},
+// 				{Name: "status", Type: "S"},
+// 				{Name: "updated_at", Type: "N"},
+// 				{Name: "is_active", Type: "BOOL"},
+// 			},
+// 		}
 
-		rendered := utils.MustParseTemplateToString(v2.UtilityFunctionsTemplate, templateMap)
-		testUtilityFunctionsContent(t, rendered)
-	})
+// 		rendered := utils.MustParseTemplateToString(v2.UtilityFunctionsTemplate, templateMap)
+// 		testUtilityFunctionsContent(t, rendered)
+// 	})
 
-	// Test with schema including SS and NS types
-	t.Run("template_with_set_types", func(t *testing.T) {
-		templateMap := v2.TemplateMap{
-			PackageName: "userprofile",
-			TableName:   "UserProfile",
-			HashKey:     "user_id",
-			RangeKey:    "profile_type",
-			AllAttributes: []common.Attribute{
-				{Name: "user_id", Type: "S"},
-				{Name: "profile_type", Type: "S"},
-				{Name: "tags", Type: "SS"},         // String Set
-				{Name: "skill_levels", Type: "NS"}, // Number Set
-				{Name: "interests", Type: "SS"},    // String Set
-				{Name: "scores", Type: "NS"},       // Number Set
-				{Name: "is_premium", Type: "BOOL"},
-				{Name: "created_at", Type: "N"},
-			},
-		}
+// 	// Test with schema including SS and NS types
+// 	t.Run("template_with_set_types", func(t *testing.T) {
+// 		templateMap := v2.TemplateMap{
+// 			PackageName: "userprofile",
+// 			TableName:   "UserProfile",
+// 			HashKey:     "user_id",
+// 			RangeKey:    "profile_type",
+// 			AllAttributes: []common.Attribute{
+// 				{Name: "user_id", Type: "S"},
+// 				{Name: "profile_type", Type: "S"},
+// 				{Name: "tags", Type: "SS"},         // String Set
+// 				{Name: "skill_levels", Type: "NS"}, // Number Set
+// 				{Name: "interests", Type: "SS"},    // String Set
+// 				{Name: "scores", Type: "NS"},       // Number Set
+// 				{Name: "is_premium", Type: "BOOL"},
+// 				{Name: "created_at", Type: "N"},
+// 			},
+// 		}
 
-		rendered := utils.MustParseTemplateToString(v2.UtilityFunctionsTemplate, templateMap)
-		testUtilityFunctionsWithSets(t, rendered)
-	})
-}
+// 		rendered := utils.MustParseTemplateToString(v2.UtilityFunctionsTemplate, templateMap)
+// 		testUtilityFunctionsWithSets(t, rendered)
+// 	})
+// }
 
-// testUtilityFunctionsContent validates the basic content of rendered utility functions template
-func testUtilityFunctionsContent(t *testing.T, rendered string) {
-	t.Helper()
+// // testUtilityFunctionsContent validates the basic content of rendered utility functions template
+// func testUtilityFunctionsContent(t *testing.T, rendered string) {
+// 	t.Helper()
 
-	// Test that the rendered code is valid Go syntax
-	// Example: parsing with imports for all used packages should succeed
-	t.Run("go_syntax_valid", func(t *testing.T) {
-		src := "package test\n\nimport (\n" +
-			"\t\"context\"\n" +
-			"\t\"encoding/json\"\n" +
-			"\t\"fmt\"\n" +
-			"\t\"strconv\"\n" +
-			"\t\"github.com/aws/aws-lambda-go/events\"\n" +
-			"\t\"github.com/aws/aws-sdk-go-v2/feature/dynamodb/attributevalue\"\n" +
-			"\t\"github.com/aws/aws-sdk-go-v2/service/dynamodb/types\"\n" +
-			")\n\n" + rendered
-		fset := token.NewFileSet()
-		_, err := parser.ParseFile(fset, "utils.go", src, parser.ParseComments)
-		require.NoError(t, err, "Rendered UtilityFunctionsTemplate should be valid Go syntax")
-	})
+// 	// Test that the rendered code is valid Go syntax
+// 	// Example: parsing with imports for all used packages should succeed
+// 	t.Run("go_syntax_valid", func(t *testing.T) {
+// 		src := "package test\n\nimport (\n" +
+// 			"\t\"context\"\n" +
+// 			"\t\"encoding/json\"\n" +
+// 			"\t\"fmt\"\n" +
+// 			"\t\"strconv\"\n" +
+// 			"\t\"github.com/aws/aws-lambda-go/events\"\n" +
+// 			"\t\"github.com/aws/aws-sdk-go-v2/feature/dynamodb/attributevalue\"\n" +
+// 			"\t\"github.com/aws/aws-sdk-go-v2/service/dynamodb/types\"\n" +
+// 			")\n\n" + rendered
+// 		fset := token.NewFileSet()
+// 		_, err := parser.ParseFile(fset, "utils.go", src, parser.ParseComments)
+// 		require.NoError(t, err, "Rendered UtilityFunctionsTemplate should be valid Go syntax")
+// 	})
 
-	// Test that all top-level helper functions are present
-	t.Run("functions_present", func(t *testing.T) {
-		funcs := []string{
-			"func BatchPutItems(",
-			"func PutItem(",
-			"func DeleteItem(",
-			"func DeleteItemFromItem(",
-			"func DeleteItemWithCondition(",
-			"func BatchDeleteItems(",
-			"func BatchDeleteItemsFromItems(",
-			"func UpdateItem(",
-			"func UpdateItemFromItem(",
-			"func UpdateItemWithCondition(",
-			"func UpdateItemWithExpression(",
-			"func IncrementAttribute(",
-			"func AddToSet(",
-			"func RemoveFromSet(",
-			"func ExtractFromDynamoDBStreamEvent(",
-			"func IsFieldModified(",
-			"func GetBoolFieldChanged(",
-			"func CreateKey(",
-			"func CreateKeyFromItem(",
-			"func CreateTriggerHandler(",
-			"func ConvertMapToAttributeValues(",
-		}
-		for _, fn := range funcs {
-			assert.Contains(t, rendered, fn, "Should contain %s", fn)
-		}
-	})
+// 	// Test that all top-level helper functions are present
+// 	t.Run("functions_present", func(t *testing.T) {
+// 		funcs := []string{
+// 			"func BatchPutItems(",
+// 			"func PutItem(",
+// 			"func DeleteItem(",
+// 			"func DeleteItemFromItem(",
+// 			"func DeleteItemWithCondition(",
+// 			"func BatchDeleteItems(",
+// 			"func BatchDeleteItemsFromItems(",
+// 			"func UpdateItem(",
+// 			"func UpdateItemFromItem(",
+// 			"func UpdateItemWithCondition(",
+// 			"func UpdateItemWithExpression(",
+// 			"func IncrementAttribute(",
+// 			"func AddToSet(",
+// 			"func RemoveFromSet(",
+// 			"func ExtractFromDynamoDBStreamEvent(",
+// 			"func IsFieldModified(",
+// 			"func GetBoolFieldChanged(",
+// 			"func CreateKey(",
+// 			"func CreateKeyFromItem(",
+// 			"func CreateTriggerHandler(",
+// 			"func ConvertMapToAttributeValues(",
+// 		}
+// 		for _, fn := range funcs {
+// 			assert.Contains(t, rendered, fn, "Should contain %s", fn)
+// 		}
+// 	})
 
-	// Test that ExtractFromDynamoDBStreamEvent has per-type branches for each attribute type
-	t.Run("extract_branches", func(t *testing.T) {
-		// String branch
-		assert.Contains(t, rendered, "val.String()", "Should handle S/String attributes")
-		// Number branch
-		assert.Contains(t, rendered, "strconv.Atoi(val.Number())", "Should handle N/Number attributes")
-		// Boolean branch
-		assert.Contains(t, rendered, "val.Boolean()", "Should handle BOOL/Boolean attributes")
-	})
+// 	// Test that ExtractFromDynamoDBStreamEvent has per-type branches for each attribute type
+// 	t.Run("extract_branches", func(t *testing.T) {
+// 		// String branch
+// 		assert.Contains(t, rendered, "val.String()", "Should handle S/String attributes")
+// 		// Number branch
+// 		assert.Contains(t, rendered, "strconv.Atoi(val.Number())", "Should handle N/Number attributes")
+// 		// Boolean branch
+// 		assert.Contains(t, rendered, "val.Boolean()", "Should handle BOOL/Boolean attributes")
+// 	})
 
-	// Test that ConvertMapToAttributeValues covers all switch cases
-	t.Run("convert_map_branches", func(t *testing.T) {
-		branches := []string{
-			"case string:",
-			"case int:", // Added support for int
-			"case float64:",
-			"case bool:",
-			"case nil:",
-			"case map[string]interface{}:",
-			"case []interface{}:",
-			"default:",
-		}
-		for _, br := range branches {
-			assert.Contains(t, rendered, br, "ConvertMapToAttributeValues should cover %s", br)
-		}
-	})
+// 	// Test that ConvertMapToAttributeValues covers all switch cases
+// 	t.Run("convert_map_branches", func(t *testing.T) {
+// 		branches := []string{
+// 			"case string:",
+// 			"case int:", // Added support for int
+// 			"case float64:",
+// 			"case bool:",
+// 			"case nil:",
+// 			"case map[string]interface{}:",
+// 			"case []interface{}:",
+// 			"default:",
+// 		}
+// 		for _, br := range branches {
+// 			assert.Contains(t, rendered, br, "ConvertMapToAttributeValues should cover %s", br)
+// 		}
+// 	})
 
-	// Test that bool values are properly handled with BOOL type
-	t.Run("bool_type_handling", func(t *testing.T) {
-		assert.Contains(t, rendered, "types.AttributeValueMemberBOOL{Value: v}", "Should use BOOL type for bool values")
-	})
+// 	// Test that bool values are properly handled with BOOL type
+// 	t.Run("bool_type_handling", func(t *testing.T) {
+// 		assert.Contains(t, rendered, "types.AttributeValueMemberBOOL{Value: v}", "Should use BOOL type for bool values")
+// 	})
 
-	// Test DeleteItem functions
-	t.Run("delete_functions", func(t *testing.T) {
-		assert.Contains(t, rendered, "dynamodb.DeleteItemInput", "Should contain DeleteItemInput")
-		assert.Contains(t, rendered, "dynamodb.BatchWriteItemInput", "Should contain BatchWriteItemInput for batch deletes")
-		assert.Contains(t, rendered, "types.WriteRequest", "Should use WriteRequest for batch operations")
-		assert.Contains(t, rendered, "DeleteRequest", "Should use DeleteRequest for delete operations")
-	})
+// 	// Test DeleteItem functions
+// 	t.Run("delete_functions", func(t *testing.T) {
+// 		assert.Contains(t, rendered, "dynamodb.DeleteItemInput", "Should contain DeleteItemInput")
+// 		assert.Contains(t, rendered, "dynamodb.BatchWriteItemInput", "Should contain BatchWriteItemInput for batch deletes")
+// 		assert.Contains(t, rendered, "types.WriteRequest", "Should use WriteRequest for batch operations")
+// 		assert.Contains(t, rendered, "DeleteRequest", "Should use DeleteRequest for delete operations")
+// 	})
 
-	// Test batch operation limits
-	t.Run("batch_limits", func(t *testing.T) {
-		assert.Contains(t, rendered, "maximum 25 items", "Should enforce DynamoDB batch limits")
-		assert.Contains(t, rendered, "len(keys) > 25", "Should check batch size limit")
-	})
+// 	// Test batch operation limits
+// 	t.Run("batch_limits", func(t *testing.T) {
+// 		assert.Contains(t, rendered, "maximum 25 items", "Should enforce DynamoDB batch limits")
+// 		assert.Contains(t, rendered, "len(keys) > 25", "Should check batch size limit")
+// 	})
 
-	// Test UpdateItem functions
-	t.Run("update_functions", func(t *testing.T) {
-		assert.Contains(t, rendered, "dynamodb.UpdateItemInput", "Should contain UpdateItemInput")
-		assert.Contains(t, rendered, "UpdateExpression", "Should use UpdateExpression")
-		assert.Contains(t, rendered, "ExpressionAttributeNames", "Should use ExpressionAttributeNames")
-		assert.Contains(t, rendered, "ExpressionAttributeValues", "Should use ExpressionAttributeValues")
-	})
+// 	// Test UpdateItem functions
+// 	t.Run("update_functions", func(t *testing.T) {
+// 		assert.Contains(t, rendered, "dynamodb.UpdateItemInput", "Should contain UpdateItemInput")
+// 		assert.Contains(t, rendered, "UpdateExpression", "Should use UpdateExpression")
+// 		assert.Contains(t, rendered, "ExpressionAttributeNames", "Should use ExpressionAttributeNames")
+// 		assert.Contains(t, rendered, "ExpressionAttributeValues", "Should use ExpressionAttributeValues")
+// 	})
 
-	// Test update expression types
-	t.Run("update_expression_types", func(t *testing.T) {
-		assert.Contains(t, rendered, "SET ", "Should support SET operations")
-		assert.Contains(t, rendered, "ADD ", "Should support ADD operations")
-		assert.Contains(t, rendered, "DELETE ", "Should support DELETE operations")
-	})
+// 	// Test update expression types
+// 	t.Run("update_expression_types", func(t *testing.T) {
+// 		assert.Contains(t, rendered, "SET ", "Should support SET operations")
+// 		assert.Contains(t, rendered, "ADD ", "Should support ADD operations")
+// 		assert.Contains(t, rendered, "DELETE ", "Should support DELETE operations")
+// 	})
 
-	// Test set operations
-	t.Run("set_operations", func(t *testing.T) {
-		assert.Contains(t, rendered, "[]string", "Should handle string sets")
-		assert.Contains(t, rendered, "[]int", "Should handle number sets")
-		assert.Contains(t, rendered, "AttributeValueMemberSS", "Should create string set values")
-		assert.Contains(t, rendered, "AttributeValueMemberNS", "Should create number set values")
-	})
+// 	// Test set operations
+// 	t.Run("set_operations", func(t *testing.T) {
+// 		assert.Contains(t, rendered, "[]string", "Should handle string sets")
+// 		assert.Contains(t, rendered, "[]int", "Should handle number sets")
+// 		assert.Contains(t, rendered, "AttributeValueMemberSS", "Should create string set values")
+// 		assert.Contains(t, rendered, "AttributeValueMemberNS", "Should create number set values")
+// 	})
 
-	// Test conditional updates
-	t.Run("conditional_updates", func(t *testing.T) {
-		assert.Contains(t, rendered, "ConditionExpression", "Should support conditional updates")
-		assert.Contains(t, rendered, "expression.UpdateBuilder", "Should support expression builder")
-		assert.Contains(t, rendered, "expression.ConditionBuilder", "Should support condition builder")
-	})
-}
+// 	// Test conditional updates
+// 	t.Run("conditional_updates", func(t *testing.T) {
+// 		assert.Contains(t, rendered, "ConditionExpression", "Should support conditional updates")
+// 		assert.Contains(t, rendered, "expression.UpdateBuilder", "Should support expression builder")
+// 		assert.Contains(t, rendered, "expression.ConditionBuilder", "Should support condition builder")
+// 	})
+// }
 
-// testUtilityFunctionsWithSets validates utility functions content with Set types
-func testUtilityFunctionsWithSets(t *testing.T, rendered string) {
-	t.Helper()
+// // testUtilityFunctionsWithSets validates utility functions content with Set types
+// func testUtilityFunctionsWithSets(t *testing.T, rendered string) {
+// 	t.Helper()
 
-	// Test that Set type branches are present in ExtractFromDynamoDBStreamEvent
-	t.Run("extract_set_branches", func(t *testing.T) {
-		// String Set branch
-		assert.Contains(t, rendered, "val.StringSet()", "Should handle SS/StringSet attributes")
-		assert.Contains(t, rendered, "if ss := val.StringSet(); ss != nil", "Should check StringSet for nil")
+// 	// Test that Set type branches are present in ExtractFromDynamoDBStreamEvent
+// 	t.Run("extract_set_branches", func(t *testing.T) {
+// 		// String Set branch
+// 		assert.Contains(t, rendered, "val.StringSet()", "Should handle SS/StringSet attributes")
+// 		assert.Contains(t, rendered, "if ss := val.StringSet(); ss != nil", "Should check StringSet for nil")
 
-		// Number Set branch
-		assert.Contains(t, rendered, "val.NumberSet()", "Should handle NS/NumberSet attributes")
-		assert.Contains(t, rendered, "if ns := val.NumberSet(); ns != nil", "Should check NumberSet for nil")
-		assert.Contains(t, rendered, "for _, numStr := range ns", "Should iterate over NumberSet")
-		assert.Contains(t, rendered, "strconv.Atoi(numStr)", "Should convert number strings to int")
-	})
+// 		// Number Set branch
+// 		assert.Contains(t, rendered, "val.NumberSet()", "Should handle NS/NumberSet attributes")
+// 		assert.Contains(t, rendered, "if ns := val.NumberSet(); ns != nil", "Should check NumberSet for nil")
+// 		assert.Contains(t, rendered, "for _, numStr := range ns", "Should iterate over NumberSet")
+// 		assert.Contains(t, rendered, "strconv.Atoi(numStr)", "Should convert number strings to int")
+// 	})
 
-	// Test that ConvertMapToAttributeValues supports Set types
-	t.Run("convert_map_set_branches", func(t *testing.T) {
-		// String Set support
-		assert.Contains(t, rendered, "case []string:", "Should handle []string for SS type")
-		assert.Contains(t, rendered, "&types.AttributeValueMemberSS{Value: v}", "Should create SS AttributeValue")
+// 	// Test that ConvertMapToAttributeValues supports Set types
+// 	t.Run("convert_map_set_branches", func(t *testing.T) {
+// 		// String Set support
+// 		assert.Contains(t, rendered, "case []string:", "Should handle []string for SS type")
+// 		assert.Contains(t, rendered, "&types.AttributeValueMemberSS{Value: v}", "Should create SS AttributeValue")
 
-		// Number Set support
-		assert.Contains(t, rendered, "case []int:", "Should handle []int for NS type")
-		assert.Contains(t, rendered, "&types.AttributeValueMemberNS{Value: numbers}", "Should create NS AttributeValue")
-		assert.Contains(t, rendered, "numbers[i] = fmt.Sprintf(\"%d\", num)", "Should convert int to string for NS")
+// 		// Number Set support
+// 		assert.Contains(t, rendered, "case []int:", "Should handle []int for NS type")
+// 		assert.Contains(t, rendered, "&types.AttributeValueMemberNS{Value: numbers}", "Should create NS AttributeValue")
+// 		assert.Contains(t, rendered, "numbers[i] = fmt.Sprintf(\"%d\", num)", "Should convert int to string for NS")
 
-		// Integer support (for single numbers)
-		assert.Contains(t, rendered, "case int:", "Should handle int type")
-		assert.Contains(t, rendered, "&types.AttributeValueMemberN{Value: fmt.Sprintf(\"%d\", v)}", "Should create N AttributeValue from int")
-	})
+// 		// Integer support (for single numbers)
+// 		assert.Contains(t, rendered, "case int:", "Should handle int type")
+// 		assert.Contains(t, rendered, "&types.AttributeValueMemberN{Value: fmt.Sprintf(\"%d\", v)}", "Should create N AttributeValue from int")
+// 	})
 
-	// Test that Set type conditions are properly generated for all SS/NS attributes
-	t.Run("set_attribute_conditions", func(t *testing.T) {
-		setAttributes := []string{
-			"tags",         // SS type
-			"skill_levels", // NS type
-			"interests",    // SS type
-			"scores",       // NS type
-		}
+// 	// Test that Set type conditions are properly generated for all SS/NS attributes
+// 	t.Run("set_attribute_conditions", func(t *testing.T) {
+// 		setAttributes := []string{
+// 			"tags",         // SS type
+// 			"skill_levels", // NS type
+// 			"interests",    // SS type
+// 			"scores",       // NS type
+// 		}
 
-		for _, attr := range setAttributes {
-			// Should have conditional blocks for each set attribute
-			expectedFieldName := utils.ToUpperCamelCase(utils.ToSafeName(attr))
-			assert.Contains(t, rendered, expectedFieldName,
-				"Should contain field name %s for attribute %s", expectedFieldName, attr)
-		}
-	})
+// 		for _, attr := range setAttributes {
+// 			// Should have conditional blocks for each set attribute
+// 			expectedFieldName := utils.ToUpperCamelCase(utils.ToSafeName(attr))
+// 			assert.Contains(t, rendered, expectedFieldName,
+// 				"Should contain field name %s for attribute %s", expectedFieldName, attr)
+// 		}
+// 	})
 
-	// Test formatting and syntax for Set type handling
-	t.Run("set_syntax_valid", func(t *testing.T) {
-		src := "package test\n\nimport (\n" +
-			"\t\"context\"\n" +
-			"\t\"encoding/json\"\n" +
-			"\t\"fmt\"\n" +
-			"\t\"strconv\"\n" +
-			"\t\"github.com/aws/aws-lambda-go/events\"\n" +
-			"\t\"github.com/aws/aws-sdk-go-v2/feature/dynamodb/attributevalue\"\n" +
-			"\t\"github.com/aws/aws-sdk-go-v2/service/dynamodb/types\"\n" +
-			")\n\n" + rendered
+// 	// Test formatting and syntax for Set type handling
+// 	t.Run("set_syntax_valid", func(t *testing.T) {
+// 		src := "package test\n\nimport (\n" +
+// 			"\t\"context\"\n" +
+// 			"\t\"encoding/json\"\n" +
+// 			"\t\"fmt\"\n" +
+// 			"\t\"strconv\"\n" +
+// 			"\t\"github.com/aws/aws-lambda-go/events\"\n" +
+// 			"\t\"github.com/aws/aws-sdk-go-v2/feature/dynamodb/attributevalue\"\n" +
+// 			"\t\"github.com/aws/aws-sdk-go-v2/service/dynamodb/types\"\n" +
+// 			")\n\n" + rendered
 
-		fset := token.NewFileSet()
-		_, err := parser.ParseFile(fset, "utils_sets.go", src, parser.ParseComments)
-		require.NoError(t, err, "Rendered template with Sets should be valid Go syntax")
-	})
-}
+// 		fset := token.NewFileSet()
+// 		_, err := parser.ParseFile(fset, "utils_sets.go", src, parser.ParseComments)
+// 		require.NoError(t, err, "Rendered template with Sets should be valid Go syntax")
+// 	})
+// }
 
-// TestUtilityFunctionsTemplateFormatting validates that the rendered template is gofmt-compliant.
-// Example: format.Source("package test\n\n" + rendered + "\n") should return no error.
-func TestUtilityFunctionsTemplateFormatting(t *testing.T) {
-	// Test formatting with basic types
-	t.Run("basic_types_formatting", func(t *testing.T) {
-		templateMap := v2.TemplateMap{
-			PackageName: "usertable",
-			TableName:   "UserTable",
-			HashKey:     "user_id",
-			RangeKey:    "created_at",
-			AllAttributes: []common.Attribute{
-				{Name: "user_id", Type: "S"},
-				{Name: "created_at", Type: "N"},
-				{Name: "status", Type: "S"},
-				{Name: "updated_at", Type: "N"},
-				{Name: "is_active", Type: "BOOL"},
-			},
-		}
+// // TestUtilityFunctionsTemplateFormatting validates that the rendered template is gofmt-compliant.
+// // Example: format.Source("package test\n\n" + rendered + "\n") should return no error.
+// func TestUtilityFunctionsTemplateFormatting(t *testing.T) {
+// 	// Test formatting with basic types
+// 	t.Run("basic_types_formatting", func(t *testing.T) {
+// 		templateMap := v2.TemplateMap{
+// 			PackageName: "usertable",
+// 			TableName:   "UserTable",
+// 			HashKey:     "user_id",
+// 			RangeKey:    "created_at",
+// 			AllAttributes: []common.Attribute{
+// 				{Name: "user_id", Type: "S"},
+// 				{Name: "created_at", Type: "N"},
+// 				{Name: "status", Type: "S"},
+// 				{Name: "updated_at", Type: "N"},
+// 				{Name: "is_active", Type: "BOOL"},
+// 			},
+// 		}
 
-		rendered := utils.MustParseTemplateToString(v2.UtilityFunctionsTemplate, templateMap)
-		testFormattingCompliance(t, rendered)
-	})
+// 		rendered := utils.MustParseTemplateToString(v2.UtilityFunctionsTemplate, templateMap)
+// 		testFormattingCompliance(t, rendered)
+// 	})
 
-	// Test formatting with Set types
-	t.Run("set_types_formatting", func(t *testing.T) {
-		templateMap := v2.TemplateMap{
-			PackageName: "userprofile",
-			TableName:   "UserProfile",
-			HashKey:     "user_id",
-			RangeKey:    "profile_type",
-			AllAttributes: []common.Attribute{
-				{Name: "user_id", Type: "S"},
-				{Name: "profile_type", Type: "S"},
-				{Name: "tags", Type: "SS"},
-				{Name: "skill_levels", Type: "NS"},
-				{Name: "interests", Type: "SS"},
-				{Name: "scores", Type: "NS"},
-				{Name: "is_premium", Type: "BOOL"},
-				{Name: "created_at", Type: "N"},
-			},
-		}
+// 	// Test formatting with Set types
+// 	t.Run("set_types_formatting", func(t *testing.T) {
+// 		templateMap := v2.TemplateMap{
+// 			PackageName: "userprofile",
+// 			TableName:   "UserProfile",
+// 			HashKey:     "user_id",
+// 			RangeKey:    "profile_type",
+// 			AllAttributes: []common.Attribute{
+// 				{Name: "user_id", Type: "S"},
+// 				{Name: "profile_type", Type: "S"},
+// 				{Name: "tags", Type: "SS"},
+// 				{Name: "skill_levels", Type: "NS"},
+// 				{Name: "interests", Type: "SS"},
+// 				{Name: "scores", Type: "NS"},
+// 				{Name: "is_premium", Type: "BOOL"},
+// 				{Name: "created_at", Type: "N"},
+// 			},
+// 		}
 
-		rendered := utils.MustParseTemplateToString(v2.UtilityFunctionsTemplate, templateMap)
-		testFormattingCompliance(t, rendered)
-	})
+// 		rendered := utils.MustParseTemplateToString(v2.UtilityFunctionsTemplate, templateMap)
+// 		testFormattingCompliance(t, rendered)
+// 	})
 
-	// Test with edge cases and complex attribute names
-	t.Run("edge_cases_formatting", func(t *testing.T) {
-		templateMap := v2.TemplateMap{
-			PackageName: "complextest",
-			TableName:   "ComplexTest",
-			HashKey:     "complex_id",
-			RangeKey:    "sort_key",
-			AllAttributes: []common.Attribute{
-				{Name: "complex_id", Type: "S"},
-				{Name: "sort_key", Type: "S"},
-				{Name: "multi_word_tags", Type: "SS"},     // Complex name
-				{Name: "skill-levels", Type: "NS"},        // With hyphen
-				{Name: "user_interests", Type: "SS"},      // Underscore
-				{Name: "test_scores_2024", Type: "NS"},    // With numbers
-				{Name: "is_active_user", Type: "BOOL"},    // Boolean
-				{Name: "created_at_timestamp", Type: "N"}, // Long name
-			},
-		}
+// 	// Test with edge cases and complex attribute names
+// 	t.Run("edge_cases_formatting", func(t *testing.T) {
+// 		templateMap := v2.TemplateMap{
+// 			PackageName: "complextest",
+// 			TableName:   "ComplexTest",
+// 			HashKey:     "complex_id",
+// 			RangeKey:    "sort_key",
+// 			AllAttributes: []common.Attribute{
+// 				{Name: "complex_id", Type: "S"},
+// 				{Name: "sort_key", Type: "S"},
+// 				{Name: "multi_word_tags", Type: "SS"},     // Complex name
+// 				{Name: "skill-levels", Type: "NS"},        // With hyphen
+// 				{Name: "user_interests", Type: "SS"},      // Underscore
+// 				{Name: "test_scores_2024", Type: "NS"},    // With numbers
+// 				{Name: "is_active_user", Type: "BOOL"},    // Boolean
+// 				{Name: "created_at_timestamp", Type: "N"}, // Long name
+// 			},
+// 		}
 
-		rendered := utils.MustParseTemplateToString(v2.UtilityFunctionsTemplate, templateMap)
-		testFormattingCompliance(t, rendered)
-	})
-}
+// 		rendered := utils.MustParseTemplateToString(v2.UtilityFunctionsTemplate, templateMap)
+// 		testFormattingCompliance(t, rendered)
+// 	})
+// }
 
-// testFormattingCompliance checks if the rendered template is gofmt-compliant
-func testFormattingCompliance(t *testing.T, rendered string) {
-	t.Helper()
+// // testFormattingCompliance checks if the rendered template is gofmt-compliant
+// func testFormattingCompliance(t *testing.T, rendered string) {
+// 	t.Helper()
 
-	full := "package test\n\nimport (\n" +
-		"\t\"context\"\n" +
-		"\t\"encoding/json\"\n" +
-		"\t\"fmt\"\n" +
-		"\t\"strconv\"\n" +
-		"\t\"github.com/aws/aws-lambda-go/events\"\n" +
-		"\t\"github.com/aws/aws-sdk-go-v2/feature/dynamodb/attributevalue\"\n" +
-		"\t\"github.com/aws/aws-sdk-go-v2/service/dynamodb/types\"\n" +
-		")\n\n" +
-		rendered + "\n"
+// 	full := "package test\n\nimport (\n" +
+// 		"\t\"context\"\n" +
+// 		"\t\"encoding/json\"\n" +
+// 		"\t\"fmt\"\n" +
+// 		"\t\"strconv\"\n" +
+// 		"\t\"github.com/aws/aws-lambda-go/events\"\n" +
+// 		"\t\"github.com/aws/aws-sdk-go-v2/feature/dynamodb/attributevalue\"\n" +
+// 		"\t\"github.com/aws/aws-sdk-go-v2/service/dynamodb/types\"\n" +
+// 		")\n\n" +
+// 		rendered + "\n"
 
-	_, err := format.Source([]byte(full))
-	require.NoError(t, err, "UtilityFunctionsTemplate should be gofmt-compliant")
-}
+// 	_, err := format.Source([]byte(full))
+// 	require.NoError(t, err, "UtilityFunctionsTemplate should be gofmt-compliant")
+// }
 
-// TestUtilityFunctionsTemplateExtractLogic validates the ExtractFromDynamoDBStreamEvent logic
-func TestUtilityFunctionsTemplateExtractLogic(t *testing.T) {
-	templateMap := v2.TemplateMap{
-		PackageName: "testextract",
-		TableName:   "TestExtract",
-		HashKey:     "id",
-		RangeKey:    "created",
-		AllAttributes: []common.Attribute{
-			{Name: "id", Type: "S"},
-			{Name: "created", Type: "N"},
-			{Name: "name", Type: "S"},
-			{Name: "tags", Type: "SS"},
-			{Name: "scores", Type: "NS"},
-			{Name: "is_active", Type: "BOOL"},
-		},
-	}
+// // TestUtilityFunctionsTemplateExtractLogic validates the ExtractFromDynamoDBStreamEvent logic
+// func TestUtilityFunctionsTemplateExtractLogic(t *testing.T) {
+// 	templateMap := v2.TemplateMap{
+// 		PackageName: "testextract",
+// 		TableName:   "TestExtract",
+// 		HashKey:     "id",
+// 		RangeKey:    "created",
+// 		AllAttributes: []common.Attribute{
+// 			{Name: "id", Type: "S"},
+// 			{Name: "created", Type: "N"},
+// 			{Name: "name", Type: "S"},
+// 			{Name: "tags", Type: "SS"},
+// 			{Name: "scores", Type: "NS"},
+// 			{Name: "is_active", Type: "BOOL"},
+// 		},
+// 	}
 
-	rendered := utils.MustParseTemplateToString(v2.UtilityFunctionsTemplate, templateMap)
+// 	rendered := utils.MustParseTemplateToString(v2.UtilityFunctionsTemplate, templateMap)
 
-	t.Run("extract_all_type_branches", func(t *testing.T) {
-		// Check that each attribute type has proper extraction logic
-		typeChecks := map[string][]string{
-			"S":    {"val.String()", "item.Id =", "item.Name ="},
-			"N":    {"strconv.Atoi(val.Number())", "item.Created ="},
-			"BOOL": {"val.Boolean()", "item.IsActive ="},
-			"SS":   {"val.StringSet()", "item.Tags ="},
-			"NS":   {"val.NumberSet()", "item.Scores ="},
-		}
+// 	t.Run("extract_all_type_branches", func(t *testing.T) {
+// 		// Check that each attribute type has proper extraction logic
+// 		typeChecks := map[string][]string{
+// 			"S":    {"val.String()", "item.Id =", "item.Name ="},
+// 			"N":    {"strconv.Atoi(val.Number())", "item.Created ="},
+// 			"BOOL": {"val.Boolean()", "item.IsActive ="},
+// 			"SS":   {"val.StringSet()", "item.Tags ="},
+// 			"NS":   {"val.NumberSet()", "item.Scores ="},
+// 		}
 
-		for typeCode, checks := range typeChecks {
-			for _, check := range checks {
-				assert.Contains(t, rendered, check,
-					"Should contain %s logic for type %s", check, typeCode)
-			}
-		}
-	})
+// 		for typeCode, checks := range typeChecks {
+// 			for _, check := range checks {
+// 				assert.Contains(t, rendered, check,
+// 					"Should contain %s logic for type %s", check, typeCode)
+// 			}
+// 		}
+// 	})
 
-	t.Run("extract_error_handling", func(t *testing.T) {
-		// Check proper error handling for number conversions
-		assert.Contains(t, rendered, "if n, err := strconv.Atoi(val.Number()); err == nil",
-			"Should have error handling for number conversion")
-		assert.Contains(t, rendered, "if num, err := strconv.Atoi(numStr); err == nil",
-			"Should have error handling for number set conversion")
-	})
+// 	t.Run("extract_error_handling", func(t *testing.T) {
+// 		// Check proper error handling for number conversions
+// 		assert.Contains(t, rendered, "if n, err := strconv.Atoi(val.Number()); err == nil",
+// 			"Should have error handling for number conversion")
+// 		assert.Contains(t, rendered, "if num, err := strconv.Atoi(numStr); err == nil",
+// 			"Should have error handling for number set conversion")
+// 	})
 
-	t.Run("extract_nil_checks", func(t *testing.T) {
-		// Check nil checks for Set types
-		assert.Contains(t, rendered, "if ss := val.StringSet(); ss != nil",
-			"Should check StringSet for nil")
-		assert.Contains(t, rendered, "if ns := val.NumberSet(); ns != nil",
-			"Should check NumberSet for nil")
-	})
-}
+// 	t.Run("extract_nil_checks", func(t *testing.T) {
+// 		// Check nil checks for Set types
+// 		assert.Contains(t, rendered, "if ss := val.StringSet(); ss != nil",
+// 			"Should check StringSet for nil")
+// 		assert.Contains(t, rendered, "if ns := val.NumberSet(); ns != nil",
+// 			"Should check NumberSet for nil")
+// 	})
+// }
